Rename package pkg/app/api/api to grpcapi
