+++ cmd/server/BUILD.bazel
@@ -11,10 +11,10 @@ go_library(
     visibility = ["//visibility:private"],
     deps = [
         "//pkg/admin:go_default_library",
-        "//pkg/app/api/api:go_default_library",
         "//pkg/app/api/applicationlivestatestore:go_default_library",
         "//pkg/app/api/authhandler:go_default_library",
         "//pkg/app/api/commandstore:go_default_library",
+        "//pkg/app/api/grpcapi:go_default_library",
         "//pkg/app/api/pipedverifier:go_default_library",
         "//pkg/app/api/service/webservice:go_default_library",
         "//pkg/app/api/stagelogstore:go_default_library",
