+++ pkg/app/api/grpcapi/deploymentconfigtemplates/KubernetesBlueGreen
