+++ pkg/app/api/grpcapi/deploymentconfigtemplates/KubernetesCanary
