+++ pkg/app/api/grpcapi/deploymentconfigtemplates/KubernetesSimple
