+++ pkg/app/api/grpcapi/piped_api_test.go
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-package api
+package grpcapi
 
 import (
 	"context"
